+++ sauron-node.js
@@ -1,3 +1,7 @@
 module.exports = {
-  extends: ["ash-nazg/node", "ash-nazg/sauron"]
+  extends: ["ash-nazg/node", "ash-nazg/sauron"],
+  rules: {
+    "node/prefer-promises/dns": ["error"],
+    "node/prefer-promises/fs": ["error"]
+  }
 };